requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
pynacl>=1.5,<2.0  # libsodium sealed-box decryption (Phase 2.5)
